Update Man_Sith_Captain.mbch
Corrected weapon to account for mando ee3 changes

Reduced Dash to level 1 and dodge to level 2 b/c this class has more firepower than the diplomat.

Changed force protect icon to the sith shield and indicated it was 'protect'

lowered defense to level 1, and changed class to Jedi so that the bp actually worked.

Removed WP_Frag since it's not in the description.