Update Man_Sith_Diplomat.mbch
Lowered armor to 30 but gave shield recharge 1 [previously didn't have any], changed description  from shield recharge 2 to 1.

Changed to Jedi to enable BP and gave defense 2. Clarified that the shields are for force absorb and made it work that way.

Fixed the classflag for fast hacking.

renamed blaster to holdout blaster.