grogu/jerec CD defaulting
Grogu has had Speed taken away and his gameplay was unproblematic with the broken default cooldown. Now that it is actually working I find it too slow and believe it is no longer needed.

Jerec uses too much FP for Destruction for the cooldown to even matter so that goes too.