v4_BX nerfs/rebalance
Dash 2 -> 1 (hybrid can still buy level 2).
Hybrid health cost changed so level 1 it's cheaper on level 1 and 3. Beskar 3 cost increased. This should mean you no longer just don't take health in favor for beskar.
BP reduced to 80/100/140 instead of 80/110/160. 160 bp for a class like that is too much as it's already quick and spammy and has guns.
I expect the bp nerf to be enough combined with the other slight tweaks. If not, then Beskar 3 should be removed and health 3/armor 3 should grant a bit more health/armor.