community class: melee mando
a melee-focused 1 life 1 slot mando with no weapons
there was some excitement for this class this summer. its pretty heavily inspired by the player, kind of a joke class with some actual fighting capacity. i tried to not just make it a flying wookie and itd be cool if this makes it past the censors. could probably use an HP/armor buff