Add files via upload
Proposed class for an analogue to Grogu for Imps but with a fresh gameplay experience by making the class an "agility healer" where the player has mobility options and force powers at their disposal but no saber. The class also has a global damage buff for their melee and offensive force powers but otherwise lacks the extra abilities of higher rank powers like Lightning Push or Grip 3. Core concept of this class is to be able to race around the map to heal and regen resources for allies but has to stay still in order to heal teammates.