[Community Class] Darth Plagueis
Classic Star Wars Legends character and Imp analogue to Grogu. Main premise of this class is a "unarmed speed healer" that instead of healing in an AoE like Gorgu or Traya, Plagueis players use their speed to run around the map aiding allies whether by restoring HP, resources, or even BP when a teammate is in trouble and losing a duel. However, stopping to help allies brings risk as they have no saber to defend themselves from the enemy team, their health is quite meager, and model is quite large. Obviously open to feedback and edits as necessary.